update:Tags Enum
