+++ backend/src/main/java/com/devu/backend/api/comment/CommentCreateRequestDto.java
@@ -1,4 +1,4 @@
-package com.devu.backend.controller.comment;
+package com.devu.backend.api.comment;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,8 +9,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
-public class CommentResponseDto {
-    private String username;
-    private String title;
-    private Long commentId;
+public class CommentCreateRequestDto {
+    private Long userId;
+    private Long postId;
+    private String contents;
 }
