+++ backend/src/main/java/com/devu/backend/entity/post/PostTags.java
@@ -1,0 +1,21 @@
+package com.devu.backend.entity.post;
+
+public enum PostTags {
+    SPRING,
+    C,
+    CPP,
+    CSHARP,
+    JAVASCRIPT,
+    REACT,
+    VUE,
+    NODE,
+    PYTHON,
+    DJANGO,
+    FLASK,
+    GO,
+    SWIFT,
+    ANGULAR,
+    RUBY,
+    JAVA,
+    FLUTTER
+}
