+++ backend/src/main/java/com/devu/backend/service/CommentService.java
@@ -4,9 +4,8 @@ import com.devu.backend.common.exception.CommentContentNullException;
 import com.devu.backend.common.exception.CommentNotFoundException;
 import com.devu.backend.common.exception.PostNotFoundException;
 import com.devu.backend.common.exception.UserNotFoundException;
-import com.devu.backend.controller.comment.CommentCreateRequestDto;
-import com.devu.backend.controller.comment.CommentResponseDto;
-import com.devu.backend.controller.comment.CommentUpdateRequestDto;
+import com.devu.backend.api.comment.CommentCreateRequestDto;
+import com.devu.backend.api.comment.CommentUpdateRequestDto;
 import com.devu.backend.entity.Comment;
 import com.devu.backend.entity.User;
 import com.devu.backend.entity.post.Post;
